PIM-3832: rename converters to array converters
